Only set range by text length for text nodes
If range.startContainer is not a text node, the second argument to
setStart acts as the number of child nodes, not the number of
characters. Setting the range to the length of the text replaced caused
a range error because it tried to start the range after "length" child
nodes.

This only sets the start of the range if startContainer is an text node.