Updates for compatibility with prosemirror
This fixes a few bugs I found when dealing with prosemirror
contenteditable divs. getCursorOffset would find the wrong bounding
rectangle for the editor. I check the rectangle for 0 values and replace
it with the parent's bounding rectangle. Also, getBeforeCursor would
return an incorrect value for the first character in a field or when all
content was deleted. I check innerText if the startContainer is not a
Text node, and only return null if that fails.