Fix issue with safari completing wrong
On Chrome, doing range.selectNode(...) followed by
document.execCommand('insertText', ...) replaces the selection
with the new text. On Safari, it does it automatically delete the
selection, so we are forced to explicitly delete it before the insert.

Since the delete was supposed to happen either way, I didn't put it
behind a browser check.